+++ wallet/api/admin_generate_api_token.go
@@ -3,6 +3,7 @@ package api
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	"code.vegaprotocol.io/vega/wallet/api/session"
@@ -11,6 +12,7 @@ import (
 
 type AdminGenerateAPITokenParams struct {
 	Description string                            `json:"name"`
+	Expiry      *int64                            `json:"expiry"`
 	Wallet      AdminGenerateAPITokenWalletParams `json:"wallet"`
 }
 
@@ -48,6 +50,7 @@ func (h *AdminGenerateAPIToken) Handle(ctx context.Context, rawParams jsonrpc.Pa
 	token := session.Token{
 		Description: params.Description,
 		Token:       session.GenerateToken(),
+		Expiry:      params.Expiry,
 		Wallet: session.WalletCredentials{
 			Name:       params.Wallet.Name,
 			Passphrase: params.Wallet.Passphrase,
@@ -79,6 +82,12 @@ func validateAdminGenerateAPITokenParams(rawParams jsonrpc.Params) (AdminGenerat
 
 	if params.Wallet.Passphrase == "" {
 		return AdminGenerateAPITokenParams{}, ErrWalletPassphraseIsRequired
+	}
+
+	if params.Expiry != nil {
+		if time.Now().After(time.Unix(*params.Expiry, 0)) {
+			return AdminGenerateAPITokenParams{}, ErrAPITokenExpiryInThePast
+		}
 	}
 
 	return params, nil