+++ wallet/api/client_sign_transaction_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"testing"
+	"time"
 
 	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	vgrand "code.vegaprotocol.io/vega/libs/rand"
@@ -23,6 +24,8 @@ func TestSignTransaction(t *testing.T) {
 	t.Run("Signing a transaction with valid params succeeds", testSigningTransactionWithValidParamsSucceeds)
 	t.Run("Signing a transaction with invalid token fails", testSigningTransactionWithInvalidTokenFails)
 	t.Run("Signing a transaction with a long-living token succeeds", testSigningTransactionWithLongLivingTokenSucceeds)
+	t.Run("Signing a transaction with a long-living token expired fails", testSigningTransactionWithLongLivingExpiredTokenFails)
+	t.Run("Signing a transaction with a long-living valid token succeeds", testSigningTransactionWithLongLivingValidTokenSucceeds)
 	t.Run("Signing a transaction without the needed permissions sign the transaction", testSigningTransactionWithoutNeededPermissionsDoesNotSignTransaction)
 	t.Run("Refusing the signing of a transaction does not sign the transaction", testRefusingSigningOfTransactionDoesNotSignTransaction)
 	t.Run("Cancelling the review does not sign the transaction", testCancellingTheReviewDoesNotSignTransaction)
@@ -114,6 +117,7 @@ func testSigningTransactionWithValidParamsSucceeds(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 	token := connectWallet(t, handler.sessions, hostname, wallet1)
 	// -- expected calls
 	handler.interactor.EXPECT().NotifyInteractionSessionBegan(ctx, gomock.Any()).Times(1).Return(nil)
@@ -153,6 +157,7 @@ func testSigningTransactionWithInvalidTokenFails(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 
 	// when
 	result, errorDetails := handler.handle(t, ctx, api.ClientSignTransactionParams{
@@ -175,7 +180,81 @@ func testSigningTransactionWithLongLivingTokenSucceeds(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
-	if err := handler.sessions.ConnectWalletForLongLivingConnection(token, wallet1); err != nil {
+	handler.time.EXPECT().Now().Times(1)
+	if err := handler.sessions.ConnectWalletForLongLivingConnection(token, wallet1, time.Now(), nil); err != nil {
+		t.Fatalf("could not connect test wallet to a long-living sessions: %v", err)
+	}
+	// -- expected calls
+	handler.interactor.EXPECT().NotifyInteractionSessionBegan(ctx, gomock.Any()).Times(1).Return(nil)
+	handler.interactor.EXPECT().NotifyInteractionSessionEnded(ctx, gomock.Any()).Times(1)
+	handler.nodeSelector.EXPECT().Node(ctx, gomock.Any()).Times(1).Return(handler.node, nil)
+	handler.node.EXPECT().LastBlock(ctx).Times(1).Return(types.LastBlock{
+		BlockHeight:             100,
+		BlockHash:               vgrand.RandomStr(64),
+		ProofOfWorkHashFunction: "sha3_24_rounds",
+		ProofOfWorkDifficulty:   1,
+		ChainID:                 vgrand.RandomStr(5),
+	}, nil)
+	handler.interactor.EXPECT().NotifySuccessfulRequest(ctx, traceID, api.TransactionSuccessfullySigned).Times(1)
+	handler.interactor.EXPECT().Log(ctx, traceID, gomock.Any(), gomock.Any()).AnyTimes()
+
+	// when
+	result, errorDetails := handler.handle(t, ctx, api.ClientSignTransactionParams{
+		Token:              token,
+		PublicKey:          kp.PublicKey(),
+		EncodedTransaction: encodedTransaction,
+	})
+
+	// then
+	assert.Nil(t, errorDetails)
+	require.NotEmpty(t, result)
+	assert.NotEmpty(t, result.Tx)
+}
+
+func testSigningTransactionWithLongLivingExpiredTokenFails(t *testing.T) {
+	// given
+	ctx, _ := contextWithTraceID()
+	encodedTransaction := "ewogICAgInZvdGVTdWJtaXNzaW9uIjogewogICAgICAgICJwcm9wb3NhbElkIjogImViMmQzOTAyZmRkYTljM2ViNmUzNjlmMjIzNTY4OWI4NzFjNzMyMmNmM2FiMjg0ZGRlM2U5ZGZjMTM4NjNhMTciLAogICAgICAgICJ2YWx1ZSI6ICJWQUxVRV9ZRVMiCiAgICB9Cn0K"
+	wallet1, kp := walletWithKey(t)
+	token := vgrand.RandomStr(10)
+
+	now := time.Now()
+	expiresAt := now.Add(1 * time.Hour)
+	nextNow := now.Add(2 * time.Hour)
+
+	// setup
+	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1).Return(nextNow)
+	if err := handler.sessions.ConnectWalletForLongLivingConnection(token, wallet1, now, &expiresAt); err != nil {
+		t.Fatalf("could not connect test wallet to a long-living sessions: %v", err)
+	}
+
+	// when
+	_, errorDetails := handler.handle(t, ctx, api.ClientSignTransactionParams{
+		Token:              token,
+		PublicKey:          kp.PublicKey(),
+		EncodedTransaction: encodedTransaction,
+	})
+
+	// then
+	assert.EqualError(t, errorDetails, "api token expired (Invalid params -32602)")
+}
+
+func testSigningTransactionWithLongLivingValidTokenSucceeds(t *testing.T) {
+	// given
+	ctx, traceID := contextWithTraceID()
+	encodedTransaction := "ewogICAgInZvdGVTdWJtaXNzaW9uIjogewogICAgICAgICJwcm9wb3NhbElkIjogImViMmQzOTAyZmRkYTljM2ViNmUzNjlmMjIzNTY4OWI4NzFjNzMyMmNmM2FiMjg0ZGRlM2U5ZGZjMTM4NjNhMTciLAogICAgICAgICJ2YWx1ZSI6ICJWQUxVRV9ZRVMiCiAgICB9Cn0K"
+	wallet1, kp := walletWithKey(t)
+	token := vgrand.RandomStr(10)
+
+	now := time.Now()
+	expiresAt := now.Add(2 * time.Hour)
+	nextNow := now.Add(1 * time.Hour)
+
+	// setup
+	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1).Return(nextNow)
+	if err := handler.sessions.ConnectWalletForLongLivingConnection(token, wallet1, now, &expiresAt); err != nil {
 		t.Fatalf("could not connect test wallet to a long-living sessions: %v", err)
 	}
 	// -- expected calls
@@ -215,6 +294,7 @@ func testSigningTransactionWithoutNeededPermissionsDoesNotSignTransaction(t *tes
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 	token := connectWallet(t, handler.sessions, hostname, wallet1)
 
 	// when
@@ -244,6 +324,7 @@ func testRefusingSigningOfTransactionDoesNotSignTransaction(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 	token := connectWallet(t, handler.sessions, hostname, wallet1)
 	// -- expected calls
 	handler.interactor.EXPECT().NotifyInteractionSessionBegan(ctx, gomock.Any()).Times(1).Return(nil)
@@ -277,6 +358,7 @@ func testCancellingTheReviewDoesNotSignTransaction(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 	token := connectWallet(t, handler.sessions, hostname, wallet1)
 	// -- expected calls
 	handler.interactor.EXPECT().NotifyInteractionSessionBegan(ctx, gomock.Any()).Times(1).Return(nil)
@@ -310,6 +392,7 @@ func testInterruptingTheRequestDoesNotSignTransaction(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 	token := connectWallet(t, handler.sessions, hostname, wallet1)
 	// -- expected calls
 	handler.interactor.EXPECT().NotifyInteractionSessionBegan(ctx, gomock.Any()).Times(1).Return(nil)
@@ -344,6 +427,7 @@ func testGettingInternalErrorDuringReviewDoesNotSignTransaction(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 	token := connectWallet(t, handler.sessions, hostname, wallet1)
 	// -- expected calls
 	handler.interactor.EXPECT().NotifyInteractionSessionBegan(ctx, gomock.Any()).Times(1).Return(nil)
@@ -378,6 +462,7 @@ func testNoHealthyNodeAvailableDoesNotSignTransaction(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 	token := connectWallet(t, handler.sessions, hostname, wallet1)
 	// -- expected calls
 	handler.interactor.EXPECT().NotifyInteractionSessionBegan(ctx, gomock.Any()).Times(1).Return(nil)
@@ -417,6 +502,7 @@ func testFailingToGetLastBlockDoesNotSignTransaction(t *testing.T) {
 
 	// setup
 	handler := newSignTransactionHandler(t)
+	handler.time.EXPECT().Now().Times(1)
 	token := connectWallet(t, handler.sessions, hostname, wallet1)
 	// -- expected calls
 	handler.interactor.EXPECT().NotifyInteractionSessionBegan(ctx, gomock.Any()).Times(1).Return(nil)
@@ -449,6 +535,7 @@ type signTransactionHandler struct {
 	sessions     *session.Sessions
 	nodeSelector *nodemock.MockSelector
 	node         *nodemock.MockNode
+	time         *mocks.MockTimeProvider
 }
 
 func (h *signTransactionHandler) handle(t *testing.T, ctx context.Context, params interface{}) (api.ClientSignTransactionResult, *jsonrpc.ErrorDetails) {
@@ -474,13 +561,15 @@ func newSignTransactionHandler(t *testing.T) *signTransactionHandler {
 
 	sessions := session.NewSessions()
 	node := nodemock.NewMockNode(ctrl)
+	tp := mocks.NewMockTimeProvider(ctrl)
 
 	return &signTransactionHandler{
-		ClientSignTransaction: api.NewSignTransaction(interactor, nodeSelector, sessions),
+		ClientSignTransaction: api.NewSignTransaction(interactor, nodeSelector, sessions, tp),
 		ctrl:                  ctrl,
 		nodeSelector:          nodeSelector,
 		interactor:            interactor,
 		sessions:              sessions,
 		node:                  node,
+		time:                  tp,
 	}
 }