+++ wallet/api/time_provider.go
@@ -1,0 +1,13 @@
+package api
+
+import "time"
+
+type TimeProvider interface {
+	Now() time.Time
+}
+
+type StdTime struct{}
+
+func (stdt *StdTime) Now() time.Time {
+	return time.Now()
+}