+++ CHANGELOG.md
@@ -30,8 +30,9 @@
 - [6871](https://github.com/vegaprotocol/vega/issues/6871) - Assure integration test framework throws an error when no watchers specified for a network parameter being set/updated
 - [6908](https://github.com/vegaprotocol/vega/issues/6871) - Update default retention policy
 - [6827](https://github.com/vegaprotocol/vega/issues/6615) - Add filters to `ordersConnection`
-- [6910](https://github.com/vegaprotocol/vega/issues/6910) - Separate settled position from position 
-- [6819](https://github.com/vegaprotocol/vega/issues/6819) -  Support long-living token in wallet client API
+- [6910](https://github.com/vegaprotocol/vega/issues/6910) - Separate settled position from position
+- [6819](https://github.com/vegaprotocol/vega/issues/6819) - Support long-living token in wallet client API
+- [6964](https://github.com/vegaprotocol/vega/issues/6964) - Add support for long living tokens with expiry
 
 
 ### 🐛 Fixes
@@ -60,7 +61,7 @@
 - [6844](https://github.com/vegaprotocol/vega/issues/6844) - Unstable `CI` tests for `dehistory`
 - [6844](https://github.com/vegaprotocol/vega/issues/6844) - Add API descriptions, remove unused ledger entries and fix typos
 - [6960](https://github.com/vegaprotocol/vega/issues/6960) - Infer has traded from settlement engine rather than from an unsaved-to-snapshot flag
- 
+
 ## 0.62.1
 
 ### 🛠 Improvements